feat: add double precision (float64) inference support
Add dtype='float64' option to MatterSimCalculator for double-precision
inference. This improves numerical stability for sensitive properties
like thermal conductivity.

Changes:
- MatterSimCalculator: new dtype parameter ('float32' or 'float64'),
  converts model weights via model.double() and creates inputs in the
  matching dtype
- m3gnet.py: replace .float() with .to(pos.dtype) for pbc_offsets so
  it respects the model's precision
- calculate(): upcast legacy-path float tensors to match model dtype

On equilibrium Si diamond, float64 reduces force noise from ~1e-6 to
~1e-15 eV/Å compared to float32.

+++ src/mattersim/forcefield/m3gnet/m3gnet.py
@@ -79,7 +79,7 @@ class M3Gnet(nn.Module):
         # Exact data from input_dictionary
         pos = input["atom_pos"]
         cell = input["cell"]
-        pbc_offsets = input["pbc_offsets"].float()
+        pbc_offsets = input["pbc_offsets"].to(pos.dtype)
         atom_attr = input["atom_attr"]
         edge_index = input["edge_index"].long()
         three_body_indices = input["three_body_indices"].long()
